GameCore: Now Using FeatureTraits for TestGame
