+++ tests/Anneck/Game/TestGameTest.php
@@ -28,7 +28,7 @@ class TestGameTest extends \PHPUnit_Framework_TestCase
     public function testGameSpecification()
     {
         $game = new TestGame();
-        self::assertEquals($game->__toString(), 'Anneck\Game\TestGame');
+        self::assertEquals('Anneck\Game\TestGame', $game->__toString());
 
         $worldConfiguration = new WorldConfiguration();
         $world = new DefaultWorld();
